feat: replase in-memory db on prisma orm for Album

+++ src/artist/artist.service.ts
@@ -1,7 +1,6 @@
 import { forwardRef, Inject, Injectable } from '@nestjs/common';
 import { Artist } from '@prisma/client';
 import { ArtistDto } from './artist.const';
-import { AlbumService } from 'src/album/album.service';
 import { TrackService } from 'src/track/track.service';
 import { FavoritesService } from 'src/favorite/favorite.service';
 import { PrismaService } from 'src/prisma/prisma.service';
@@ -10,53 +9,55 @@ import { PrismaService } from 'src/prisma/prisma.service';
 export class ArtistService {
   constructor(
     private readonly prisma: PrismaService,
-    @Inject(forwardRef(() => AlbumService))
-    private readonly albumService: AlbumService,
     @Inject(forwardRef(() => TrackService))
     private readonly trackService: TrackService,
     @Inject(forwardRef(() => FavoritesService))
     private readonly favouritesService: FavoritesService,
   ) {}
 
+  private _artists = this.prisma.client.artist;
+
   public async getAll(): Promise<Artist[]> {
-    return await this.prisma.client.artist.findMany();
+    return await this._artists.findMany();
   }
 
   public async getById(id: string): Promise<Artist | null> {
-    return await this.prisma.client.artist.findUnique({
+    return await this._artists.findUnique({
       where: { id },
     });
   }
 
   public async create(createArtistDto: ArtistDto): Promise<Artist> {
-    return await this.prisma.client.artist.create({
+    return await this._artists.create({
       data: createArtistDto,
     });
   }
 
   public async delete(id: string): Promise<boolean> {
-    try {
-      await this.prisma.client.artist.delete({
-        where: { id },
-      });
-      this.albumService.deleteArtistId(id);
-      this.trackService.deleteArtistId(id);
-      this.favouritesService.deleteArtist(id);
-      return true;
-    } catch {
-      return false;
-    }
+    const artist = await this._artists.findUnique({
+      where: { id },
+    });
+    if (!artist) return false;
+    await this._artists.delete({
+      where: { id },
+    });
+    this.trackService.deleteArtistId(id);
+    this.favouritesService.deleteArtist(id);
+    return true;
   }
 
   public async update({
     id,
     ...updateArtistDto
   }: Artist): Promise<Artist | null> {
-    const updatedArtist = await this.prisma.client.artist.update({
+    const artist = await this._artists.findUnique({
+      where: { id },
+    });
+    if (!artist) return null;
+    const updatedArtist = await this._artists.update({
       where: { id },
       data: updateArtistDto,
     });
-    if (!updatedArtist) return null;
     return updatedArtist;
   }
 }
